add hitbox

+++ src/utilz/helpMethods.java
@@ -3,7 +3,7 @@ package utilz;
 import Main.Game;
 
 public class helpMethods {
-    public static boolean CanMoveHere (float x, float y, int width, int height, int [][] lvlData){
+    public static boolean CanMoveHere (float x, float y, float width, float height, int [][] lvlData){
         if (!isSolid(x,y, lvlData)){
             if (!isSolid(x+width, y+height, lvlData))
                 if (!isSolid(x+width, y, lvlData))
